# Conflicts:
#	Capstone Project/Assets/Art
Assets/Backgrounds/V2/En1_Day1_Material.mat
#	Capstone Project/Assets/Prefabs/Enemies/Charger_Enemy.meta
#	Capstone Project/Assets/Prefabs/Enemies/Sniper_Enemy.meta

+++ Capstone Project/Assets/Scripts/Main Player/PlayerAnimations.cs
@@ -78,7 +78,7 @@ public class PlayerAnimations : MonoBehaviour {
     }
 
     private void Play(int index) {
-        _spriterAnimator.Play(_animationList[(int)index]);
+        _spriterAnimator.Play(_animationList[index]);
     }
 }
